Move DependencyTransformation tests to new file
